+++ backend/src/test/graphql/usermanagement.test.ts
@@ -48,7 +48,8 @@ describe('Testing usermanagement', () => {
 
 	beforeEach(async () => {
 
-		await User.deleteMany({});
+		//await User.deleteMany({});
+		await User.deleteMany( {  "username" : { $ne : "adminUser" } } );
 
 		const testuser: UserSchemaType = {
 			username: 'usernameTest',
@@ -184,7 +185,8 @@ describe('Testing usermanagement', () => {
 
 
 	afterAll(async () => {
-		await User.deleteMany({});
+		//await User.deleteMany({});
+		await User.deleteMany( {  "username" : { $ne : "adminUser" } } );
 		void mongoose.connection.close();
 		console.log('Database connection closed');
 	});
